requirements.txt
selenium>=4.15.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
Office365-REST-Python-Client>=2.5.0